+++ birdset/configs/__init__.py
@@ -1,3 +1,3 @@
 from .datamodule_configs import DatasetConfig, LoadersConfig, LoaderConfig
 from .model_configs import PretrainInfoConfig
-from .module_configs import MetricsConfig, LRSchedulerConfig, LoggingParamsConfig, NetworkConfig
+from .module_configs import LRSchedulerConfig, LoggingParamsConfig, NetworkConfig, MulticlassMetricsConfig, MultilabelMetricsConfig, MultilabelMetricsConfig as MetricsConfig
